# tests/contracts/openapi.json
{
 "components": {
  "schemas": {
   "ActiveContextRequest": {
    "properties": {
     "workflow_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ]
     }
    },
    "type": "object"
   },
   "ActiveContextResponse": {
    "properties": {
     "workflow_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ]
     }
    },
    "type": "object"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array"
     }
    },
    "type": "object"
   },
...
   "ValidationError": {
    "properties": {
     "ctx": {
      "type": "object"
     },
     "input": {},
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array"
     },
     "msg": {
      "type": "string"
     },
     "type": {
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "type": "object"
   },
...
  }
 },
 "info": {
  "version": "snapshot"
 },
 "openapi": "3.1.0",
 "paths": {
  "/api/ai/active-context": {
   "post": {
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ActiveContextRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ActiveContextResponse"
        }
       }
      }
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "tags": [
     "ai"
    ]
   }
  },
  "/api/ai/availability": {
   "get": {
    "parameters": [
     {
      "in": "query",
      "name": "refresh",
      "required": false,
      "schema": {
       "default": false,
       "type": "boolean"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object"
        }
       }
      }
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "tags": [
     "ai"
    ]
   }
  },
...
 }
}